moved server related code to separate file
